Crate substrate_consensus_babe

BABE consensus

BABE (Blind Assignment for Blockchain Extension) consensus in Substrate.

Stability

This crate is highly unstable and experimental. Breaking changes may happen at any point. This crate is also missing features, such as banning of malicious validators, that are essential for a production network.

Structs

BabeConfiguration
Configuration data used by the BABE consensus engine.

BabeParams
Parameters for BABE.

BabePreDigest
A BABE pre-digest. It includes:

BabeVerifier
A verifier for Babe blocks.

Config
A slot duration. Create with get_or_compute.

Constants

BABE_ENGINE_ID
The ConsensusEngineId of BABE.

BABE_VRF_PREFIX
The prefix used by BABE for its VRF keys.

Traits

BabeApi
API necessary for block authorship with BABE.

SyncOracle
An oracle for when major synchronization work is being undertaken.

Functions

start_babe
Start the babe worker. The returned future should be run in a tokio runtime.

Type Definitions

AuthorityId
A Babe authority identifier. Necessarily equivalent to the schnorrkel public key used in the main Babe module. If that ever changes, then this must, too.

BabeImportQueue
The BABE import queue type.